move queries to module
