+++ lib/db/schema/location.ts
@@ -38,3 +38,5 @@ export const locationInsertSchema = z.object({
   lat: z.number().min(-90).max(90),
   long: z.number().min(-180).max(180),
 });
+
+export type LocationInsertSchema = z.infer<typeof locationInsertSchema>;
